door_alarm_system/requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-WTF==1.2.2
Flask-SocketIO==5.3.6
WTForms==3.0.1
eventlet==0.33.3
RPi.GPIO==0.7.1
python-dotenv==1.0.0
reportlab==4.0.4
pandas==2.0.3
pygame
pytz
requests
python-socketio
opencv-python-headless